refactor(services): extract domain-access helpers to services/domain_access.py — fix backward dep
``certificate_service.py`` and ``quiz_service.py`` were importing
``assert_course_owner`` and ``resolve_chapter_course_id`` from
``app.api.dependencies``. That's a backwards arrow (service → api)
that violates the layer hierarchy (api → service → model) and made
the two services harder to unit-test in isolation.

Move both functions to a new ``app.services.domain_access`` module.
Both are pure domain logic — one in-memory ownership check, one
single-query lookup; neither is registered with FastAPI's
``Depends(...)``, so they belong in services anyway.

``app.api.dependencies`` re-exports both names so existing route
code that imports from there keeps working without churn. The
canonical import path going forward is
``from app.services.domain_access import ...``.

Discovered by the deep architecture audit on 2026-05-25. Verifies:

  - ``python -m pytest tests/`` — 756 passed
  - ``python -m ruff check .`` clean
  - ``mypy --config-file mypy.ini`` — 118 source files, no issues

Config-centralization (Datadog / VERCEL flag / YOUVERSION_API_KEY
direct os.environ reads) is intentionally NOT in this PR — scope
discipline. It'll land as a separate small follow-up.
